# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Intelligence & Question Extraction Service",
    "description": "Asynchronous document processing service for question paper and answer key extraction.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/{id}/questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "List extracted questions for a document with filtering and sorting",
        "description": "Returns paginated questions extracted from a document.\n\nRaises 409 DOCUMENT_NOT_READY if document is currently queued or processing.",
        "operationId": "get_document_questions_api_v1_documents__id__questions_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by question status",
              "title": "Status"
            },
            "description": "Filter by question status"
          },
          {
            "name": "needs_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter questions needing human review",
              "title": "Needs Review"
            },
            "description": "Filter questions needing human review"
          },
          {
            "name": "answer_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by answer status",
              "title": "Answer Status"
            },
            "description": "Filter by answer status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by page number",
              "title": "Page"
            },
            "description": "Filter by page number"
          },
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by section heading",
              "title": "Section"
            },
            "description": "Filter by section heading"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by question type",
              "title": "Type"
            },
            "description": "Filter by question type"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum confidence",
              "title": "Min Confidence"
            },
            "description": "Minimum confidence"
          },
          {
            "name": "max_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum confidence",
              "title": "Max Confidence"
            },
            "description": "Maximum confidence"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sequence|confidence)$",
              "description": "Sort order",
              "default": "sequence",
              "title": "Sort"
            },
            "description": "Sort order"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{id}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get full details of a specific question",
        "description": "Returns full details of an individual question with review state.",
        "operationId": "get_question_api_v1_questions__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Apply human correction to a question",
        "description": "Applies human corrections to question and records an audit revision.",
        "operationId": "update_question_api_v1_questions__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{id}/questions/{qid}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get question by document ID and question ID",
        "description": "Returns question details ensuring it belongs to the specified document.",
        "operationId": "get_document_question_by_id_api_v1_documents__id__questions__qid__get",
        "parameters": [
          {
            "name": "qid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Qid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{id}/review": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Submit a human review decision for a question",
        "description": "Approve or reject an extracted question during human review.",
        "operationId": "submit_question_review_api_v1_questions__id__review_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReviewActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{id}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Stream question asset image (figure / table)",
        "description": "Streams cropped figure or table image for a question.",
        "operationId": "get_asset_api_v1_assets__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionAnswerSchema": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "not_found"
          },
          "value": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Value"
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "source": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "title": "QuestionAnswerSchema"
      },
      "QuestionAssetSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "table_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Markdown"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "page",
          "url"
        ],
        "title": "QuestionAssetSchema"
      },
...
      "QuestionFlagSchema": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "severity"
        ],
        "title": "QuestionFlagSchema"
      },
      "QuestionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "QuestionListResponse"
      },
      "QuestionNumberSchema": {
        "properties": {
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "normalized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized"
          },
          "inferred": {
            "type": "boolean",
            "title": "Inferred",
            "default": false
          }
        },
        "type": "object",
        "title": "QuestionNumberSchema"
      },
      "QuestionOptionSchema": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "raw_label": {
            "type": "string",
            "title": "Raw Label"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "label",
          "raw_label",
          "text"
        ],
        "title": "QuestionOptionSchema"
      },
      "QuestionPatchRequest": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "number_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Raw"
          },
          "number_norm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Norm"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuestionOptionSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "answer_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Raw"
          },
          "answer_value": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Value"
          },
          "answer_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Status"
          }
        },
        "type": "object",
        "title": "QuestionPatchRequest"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "number": {
            "$ref": "#/components/schemas/QuestionNumberSchema"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/QuestionOptionSchema"
            },
            "type": "array",
            "title": "Options"
          },
          "answer": {
            "$ref": "#/components/schemas/QuestionAnswerSchema"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/QuestionAssetSchema"
            },
            "type": "array",
            "title": "Assets"
          },
          "source": {
            "$ref": "#/components/schemas/QuestionSourceSchema"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/QuestionFlagSchema"
            },
            "type": "array",
            "title": "Flags"
          },
          "review": {
            "$ref": "#/components/schemas/QuestionReviewSchema"
          },
          "extraction": {
            "$ref": "#/components/schemas/QuestionExtractionSchema"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "sequence",
          "number",
          "type",
          "text",
          "answer",
          "source",
          "confidence",
          "status",
          "review",
          "extraction",
          "created_at"
        ],
        "title": "QuestionResponse"
      },
      "QuestionReviewActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(approve|reject|pending)$",
            "title": "Action"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "QuestionReviewActionRequest"
      },
...
      "QuestionSourceSchema": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pages"
          },
          "bbox_by_page": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bbox By Page"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "QuestionSourceSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}